Manually lint src/main
