+++ src/main/java/gov/nasa/podaac/swodlr/SwodlrApp.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class SwodlrApp {
-    public static void main(String[] args) {
-        SpringApplication.run(SwodlrApp.class, args);
-    }
+  public static void main(String[] args) {
+    SpringApplication.run(SwodlrApp.class, args);
+  }
 }
