+++ src/main/java/gov/nasa/podaac/swodlr/Utils.java
@@ -3,5 +3,5 @@ package gov.nasa.podaac.swodlr;
 import java.util.UUID;
 
 public class Utils {
-    public static final UUID NULL_UUID = UUID.fromString("00000000-0000-0000-0000-000000000000");
+  public static final UUID NULL_UUID = UUID.fromString("00000000-0000-0000-0000-000000000000");
 }
