+++ src/main/java/gov/nasa/podaac/swodlr/exception/SwodlrException.java
@@ -1,7 +1,7 @@
 package gov.nasa.podaac.swodlr.exception;
 
 public class SwodlrException extends RuntimeException {
-    public SwodlrException(String message) {
-        super(message);
-    }
+  public SwodlrException(String message) {
+    super(message);
+  }
 }
