+++ src/main/java/gov/nasa/podaac/swodlr/exception/SwodlrExceptionResolver.java
@@ -1,19 +1,19 @@
 package gov.nasa.podaac.swodlr.exception;
 
-import org.springframework.graphql.execution.DataFetcherExceptionResolverAdapter;
-import org.springframework.stereotype.Component;
-
 import graphql.GraphQLError;
 import graphql.GraphqlErrorBuilder;
 import graphql.schema.DataFetchingEnvironment;
+import org.springframework.graphql.execution.DataFetcherExceptionResolverAdapter;
+import org.springframework.stereotype.Component;
 
 @Component
 public class SwodlrExceptionResolver extends DataFetcherExceptionResolverAdapter {
-    @Override
-    public GraphQLError resolveToSingleError(Throwable ex, DataFetchingEnvironment env) {
-        if (!(ex instanceof SwodlrException))
-		    return null;
+  @Override
+  public GraphQLError resolveToSingleError(Throwable ex, DataFetchingEnvironment env) {
+    if (!(ex instanceof SwodlrException)) {
+      return null;
+    }
 
-        return GraphqlErrorBuilder.newError(env).message(ex.getMessage()).build();
-	}
+    return GraphqlErrorBuilder.newError(env).message(ex.getMessage()).build();
+  }
 }
