+++ src/main/java/gov/nasa/podaac/swodlr/rasterdefinition/RasterDefinitionRepository.java
@@ -1,10 +1,9 @@
-package gov.nasa.podaac.swodlr.raster_definition;
+package gov.nasa.podaac.swodlr.rasterdefinition;
 
 import java.util.List;
 import java.util.UUID;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RasterDefinitionRepository extends JpaRepository<RasterDefinition, UUID> {
-    List<RasterDefinition> findById(RasterDefinition definition);
+  List<RasterDefinition> findById(RasterDefinition definition);
 }
