+++ src/main/java/gov/nasa/podaac/swodlr/security/AuthenticationInterceptor.java
@@ -1,8 +1,8 @@
 package gov.nasa.podaac.swodlr.security;
 
+import gov.nasa.podaac.swodlr.user.UserRepository;
 import java.util.Collections;
 import java.util.UUID;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.server.WebGraphQlInterceptor;
 import org.springframework.graphql.server.WebGraphQlRequest;
@@ -10,25 +10,24 @@ import org.springframework.graphql.server.WebGraphQlResponse;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 
-import gov.nasa.podaac.swodlr.user.UserRepository;
-
 @Component
 public class AuthenticationInterceptor implements WebGraphQlInterceptor {
-    private final UUID TEMP_USER_ID = UUID.fromString("fee1dc78-0604-4fa6-adae-0b4b55440e7d");
+  private static final UUID TEMP_USER_ID = UUID.fromString("fee1dc78-0604-4fa6-adae-0b4b55440e7d");
 
-    @Autowired
-    private UserRepository userRepository;
+  @Autowired
+  private UserRepository userRepository;
 
-    @Override
-    public Mono<WebGraphQlResponse> intercept(WebGraphQlRequest request, Chain chain) {
-        request.configureExecutionInput((executionInput, builder) -> {
-            var user = userRepository.findById(TEMP_USER_ID);
+  @Override
+  public Mono<WebGraphQlResponse> intercept(WebGraphQlRequest request, Chain chain) {
+    request.configureExecutionInput((executionInput, builder) -> {
+      var user = userRepository.findById(TEMP_USER_ID);
 
-            if (user.isPresent())
-                builder.graphQLContext(Collections.singletonMap("user", user.get()));
+      if (user.isPresent()) {
+        builder.graphQLContext(Collections.singletonMap("user", user.get()));
+      }
 
-            return builder.build();
-        });
-        return chain.next(request);
-    }
+      return builder.build();
+    });
+    return chain.next(request);
+  }
 }
