+++ src/main/java/gov/nasa/podaac/swodlr/user/UserRepository.java
@@ -1,7 +1,6 @@
 package gov.nasa.podaac.swodlr.user;
 
 import java.util.UUID;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, UUID> {
